+++ mne/io/fiff/raw.py
@@ -34,11 +34,15 @@ class Raw(BaseRaw):
 
     Parameters
     ----------
-    fname : str
-        The raw file to load. For files that have automatically been split,
+    fname : str | file-like
+        The raw filename to load. For files that have automatically been split,
         the split part will be automatically loaded. Filenames should end
         with raw.fif, raw.fif.gz, raw_sss.fif, raw_sss.fif.gz,
-        raw_tsss.fif or raw_tsss.fif.gz.
+        raw_tsss.fif or raw_tsss.fif.gz. If a file-like object is provided,
+        preloading must be used.
+
+        .. versionchanged:: 0.18
+           Support for file-like objects.
     allow_maxshield : bool | str (default False)
         If True, allow loading of data that has been recorded with internal
         active compensation (MaxShield). Data recorded with MaxShield should
@@ -73,24 +77,23 @@ class Raw(BaseRaw):
     @verbose
     def __init__(self, fname, allow_maxshield=False, preload=False,
                  verbose=None):  # noqa: D102
-        fnames = [op.realpath(fname)]
-        split_fnames = []
-
         raws = []
-        for ii, this_fname in enumerate(fnames):
-            do_check_fname = this_fname not in split_fnames
+        do_check_fname = isinstance(fname, str)
+        next_fname = fname
+        while next_fname is not None:
             raw, next_fname, buffer_size_sec = \
-                self._read_raw_file(this_fname, allow_maxshield,
+                self._read_raw_file(next_fname, allow_maxshield,
                                     preload, do_check_fname)
+            do_check_fname = False
             raws.append(raw)
             if next_fname is not None:
                 if not op.exists(next_fname):
                     warn('Split raw file detected but next file %s does not '
                          'exist.' % next_fname)
-                    continue
-                # process this file next
-                fnames.insert(ii + 1, next_fname)
-                split_fnames.append(next_fname)
+                    break
+        if not isinstance(fname, str):
+            # avoid serialization error when copying file-like
+            fname = None  # noqa
 
         _check_raw_compatibility(raws)
         super(Raw, self).__init__(
@@ -129,6 +132,9 @@ class Raw(BaseRaw):
             self._preload_data(preload)
         else:
             self.preload = False
+        # If using a file-like object, fix the filenames to be representative
+        # strings now instead of the file-like objects
+        self._filenames = [_get_fname_rep(fname) for fname in self._filenames]
 
     @verbose
     def _read_raw_file(self, fname, allow_maxshield, preload,
@@ -142,8 +148,18 @@ class Raw(BaseRaw):
                                        'raw_sss.fif.gz', 'raw_tsss.fif.gz'))
 
         #   Read in the whole file if preload is on and .fif.gz (saves time)
-        ext = os.path.splitext(fname)[1].lower()
-        whole_file = preload if '.gz' in ext else False
+        if isinstance(fname, str):
+            # filename
+            fname = op.realpath(fname)
+            ext = os.path.splitext(fname)[1].lower()
+            whole_file = preload if '.gz' in ext else False
+            del ext
+        else:
+            # file-like
+            if not preload:
+                raise ValueError('preload must be used with file-like objects')
+            whole_file = True
+        fname_rep = _get_fname_rep(fname)
         ff, tree, _ = fiff_open(fname, preload=whole_file)
         with ff as fid:
             #   Read the measurement info
@@ -158,7 +174,7 @@ class Raw(BaseRaw):
                 if (len(raw_node) == 0):
                     raw_node = dir_tree_find(meas, FIFF.FIFFB_SMSH_RAW_DATA)
                     if (len(raw_node) == 0):
-                        raise ValueError('No raw data in %s' % fname)
+                        raise ValueError('No raw data in %s' % fname_rep)
                     _check_maxshield(allow_maxshield)
                     info['maxshield'] = True
 
@@ -260,7 +276,7 @@ class Raw(BaseRaw):
                                            nsamp=nsamp))
                     first_samp += nsamp
 
-            next_fname = _get_next_fname(fid, fname, tree)
+            next_fname = _get_next_fname(fid, fname_rep, tree)
 
         raw.last_samp = first_samp - 1
         raw.orig_format = orig_format
@@ -409,6 +425,13 @@ class Raw(BaseRaw):
         return self._acqparser
 
 
+def _get_fname_rep(fname):
+    if isinstance(fname, str):
+        return fname
+    else:
+        return 'File-like %r' % (fname,)
+
+
 def _check_entry(first, nent):
     """Sanity check entries."""
     if first >= nent:
@@ -421,11 +444,15 @@ def read_raw_fif(fname, allow_maxshield=False, preload=False, verbose=None):
 
     Parameters
     ----------
-    fname : str
-        The raw file to load. For files that have automatically been split,
+    fname : str | file-like
+        The raw filename to load. For files that have automatically been split,
         the split part will be automatically loaded. Filenames should end
         with raw.fif, raw.fif.gz, raw_sss.fif, raw_sss.fif.gz,
-        raw_tsss.fif or raw_tsss.fif.gz.
+        raw_tsss.fif or raw_tsss.fif.gz. If a file-like object is provided,
+        preloading must be used.
+
+        .. versionchanged:: 0.18
+           Support for file-like objects.
     allow_maxshield : bool | str (default False)
         If True, allow loading of data that has been recorded with internal
         active compensation (MaxShield). Data recorded with MaxShield should